Move memory headers into new directory
