+++ include/hip_buffer_util.hpp
@@ -6,7 +6,7 @@
 #ifndef HIP_BUFFER_UTIL_HPP
 #define HIP_BUFFER_UTIL_HPP
 
-#include "hip_recycling_allocators.hpp"
+#include "/cppuddle/memory_recycling/hip_recycling_allocators.hpp"
 
 namespace recycler {
