+++ include/stream_manager.hpp
@@ -17,7 +17,7 @@
 #include <tuple>
 #include <type_traits>
 
-#include "../include/detail/config.hpp"
+#include "cppuddle/common/config.hpp"
 
 // Need to cuda/hip definitions for default params when NOT
 // drawing from an executor pool
@@ -39,6 +39,10 @@ enum class execution_space_mode { global, independent };
 #endif
 #endif
 
+/* namespace cppuddle { */
+/* namespace executor_recycling { */
+
+namespace detail {
 /// Turns a std::array_mutex into an scoped lock
 template<typename mutex_array_t>
 auto make_scoped_lock_from_array(mutex_array_t& mutexes)
@@ -46,6 +50,7 @@ auto make_scoped_lock_from_array(mutex_array_t& mutexes)
     return std::apply([](auto&... mutexes) { return std::scoped_lock{mutexes...}; }, 
                       mutexes);
 }
+} // namespace detail
 
 template <class Interface> class round_robin_pool {
 private:
@@ -197,7 +202,7 @@ private:
     static void init(size_t number_of_streams, Ts ... executor_args) {
       /* static_assert(sizeof...(Ts) == sizeof...(Ts) && cppuddle::max_number_gpus == 1, */
       /*               "deprecated stream_pool::init does not support multigpu"); */
-      auto guard = make_scoped_lock_from_array(instance().gpu_mutexes);
+      auto guard = detail::make_scoped_lock_from_array(instance().gpu_mutexes);
       instance().streampools.emplace_back(number_of_streams, executor_args...);
       assert(instance().streampools.size() <= cppuddle::max_number_gpus);
     }
@@ -205,7 +210,7 @@ private:
     /// Multi-GPU init where executors / interfaces on all GPUs are initialized with the same arguments
     template <typename... Ts>
     static void init_all_executor_pools(size_t number_of_streams, Ts ... executor_args) {
-      auto guard = make_scoped_lock_from_array(instance().gpu_mutexes);
+      auto guard = detail::make_scoped_lock_from_array(instance().gpu_mutexes);
       if (number_of_streams > 0) {
         for (size_t gpu_id = 0; gpu_id < cppuddle::max_number_gpus; gpu_id++) {
           instance().select_gpu_function(gpu_id);
@@ -220,7 +225,7 @@ private:
     /// (useful for executor that expect an GPU-id during construction)
     template <typename... Ts>
     static void init_executor_pool(size_t gpu_id, size_t number_of_streams, Ts ... executor_args) {
-      auto guard = make_scoped_lock_from_array(instance().gpu_mutexes);
+      auto guard = detail::make_scoped_lock_from_array(instance().gpu_mutexes);
       if (number_of_streams > 0) {
         instance().select_gpu_function(gpu_id);
         instance().streampools.emplace_back(number_of_streams, 
@@ -231,7 +236,7 @@ private:
 
     // TODO add/rename into finalize?
     static void cleanup() {
-      auto guard = make_scoped_lock_from_array(instance().gpu_mutexes);
+      auto guard = detail::make_scoped_lock_from_array(instance().gpu_mutexes);
       assert(instance().streampools.size() == cppuddle::max_number_gpus);
       instance().streampools.clear();
     }
@@ -264,7 +269,7 @@ private:
     /* } */
 
     static void set_device_selector(std::function<void(size_t)> select_gpu_function) {
-      auto guard = make_scoped_lock_from_array(instance().gpu_mutexes);
+      auto guard = detail::make_scoped_lock_from_array(instance().gpu_mutexes);
       instance().select_gpu_function = select_gpu_function;
     }
 
@@ -409,5 +414,8 @@ public:
   Interface &interface;
 };
 #endif
+
+/* } // namespace executor_recycling */
+/* } // namespace cppuddle */
 
 #endif
