+++ tests/allocator_kokkos_executor_for_loop_test.cpp
@@ -21,9 +21,9 @@
 #include <hpx/timing/high_resolution_timer.hpp>
 #include <memory>
 
-#include "std_recycling_allocators.hpp"
-#include "cuda_recycling_allocators.hpp"
-#include "recycling_kokkos_view.hpp"
+#include "cppuddle/memory_recycling/std_recycling_allocators.hpp"
+#include "cppuddle/memory_recycling/cuda_recycling_allocators.hpp"
+#include "cppuddle/memory_recycling/recycling_kokkos_view.hpp"
 
 // Assert during Release builds as well for this file:
 #undef NDEBUG
